requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
streamlit>=1.30.0
sqlalchemy>=2.0.25
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.6
loguru>=0.7.2
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.26.0 